pyproject.toml
[tool.poetry]
name = "brc-website"
version = "0.1.0"
description = ""
authors = ["Ciran GÜRBÜZ <ciran.gurbuz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.0"
pytest = "^7.4.3"
black = "^23.11.0"
flake8 = "^6.1.0"
flask-wtf = "^1.2.1"
python-dotenv = "^1.0.0"
flask-mail = "^0.9.1"
pytest-cov = "^4.1.0"
pytest-flask = "^1.3.0"
pytest-mock = "^3.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
